requirements.txt
# AI Pentesting Playground
# Python dependencies

# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & HTTP
httpx==0.26.0
aiofiles==23.2.1
websockets==12.0

# YAML & Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Browser Automation
playwright==1.40.0

# System & Process
psutil==5.9.6

# OpenAI client (for Ollama compatibility)
openai==1.12.0
